Переносит логику отображения карточек в контроллер
